+++ modules/ingress/mqtt_subscriber_nodejs/src/services/SchemaRegistryService.ts
@@ -1,4 +1,4 @@
-import { getSchemaRegistryConfig, SchemaRegistryConfig } from "../config";
+import { getSchemaRegistryConfig, SchemaRegistryConfig } from "../utils/config";
 import { logger } from "../utils/logger";
 
 interface SchemaVersion {
